+++ cmn-data-svc-server/src/main/java/org/acumos/cds/repository/ThreadRepository.java
@@ -21,8 +21,23 @@
 package org.acumos.cds.repository;
 
 import org.acumos.cds.domain.MLPThread;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 public interface ThreadRepository extends PagingAndSortingRepository<MLPThread, String> {
+
+	/**
+	 * Gets a page of threads with the specified IDs using Spring magic.
+	 * 
+	 * @param solutionId
+	 *            Solution ID
+	 * @param revisionId
+	 *            Revision ID
+	 * @param pageable
+	 *            Page and sort criteria
+	 * @return Page of MLPComment
+	 */
+	Page<MLPThread> findBySolutionIdAndRevisionId(String solutionId, String revisionId, Pageable pageable);
 
 }