+++ docs/release-notes.rst
@@ -5,15 +5,25 @@ Release Notes
 The client and server are released together.  The client is deployed as a jar file to a 
 Maven Nexus repository. The server is deployed as a Docker image to a Docker registry.  
 
+Version 1.11, 3 January 2018
+----------------------------
+
+* Revise MLPSiteConfig to make userId optional
+* Revise MLPThread to add solutionId and revisionId; drop url
+* Revise MLPComment to drop url
+* Revise MLPPeer to add trustLevel
+* Add methods to query for threads and comments using solution and revision IDs 
+* Requires database schema version 1.11
+
 Version 1.10.2, 20 December 2017
--------------------------------
+--------------------------------
 
 * Extend MLPSolution with tags and solution web statistics via unidirectional annotations 
 * Extend the find-solutions method for Portal/Marketplace dynamic search
 * Requires database schema version 1.10
 
 Version 1.10.1, 12 December 2017
--------------------------------
+--------------------------------
 
 * Revert search-solutions method to version of 1.9.0
 * New find-solutions method for Portal/Marketplace dynamic search